+++ lib/rubocop/cop/style/multiline_method_signature.rb
@@ -1,0 +1,65 @@
+# frozen_string_literal: true
+
+module RuboCop
+  module Cop
+    module Style
+      # This cop checks for method signatures that span multiple lines.
+      #
+      # @example
+      #
+      #   # good
+      #
+      #   def foo(bar, baz)
+      #   end
+      #
+      #   # bad
+      #
+      #   def foo(bar,
+      #           baz)
+      #   end
+      #
+      class MultilineMethodSignature < Cop
+        MSG = 'Avoid multi-line method signatures.'.freeze
+
+        def on_def(node)
+          return if opening_line(node) == closing_line(node)
+          return if correction_exceeds_max_line_length?(node)
+
+          add_offense(node)
+        end
+        alias on_defs on_def
+
+        private
+
+        def opening_line(node)
+          node.first_line
+        end
+
+        def closing_line(node)
+          if node.arguments?
+            node.arguments.last_line
+          else
+            node.first_line
+          end
+        end
+
+        def correction_exceeds_max_line_length?(node)
+          indentation_width(node) + definition_width(node) > max_line_length
+        end
+
+        def indentation_width(node)
+          processed_source.line_indentation(node.loc.expression.line)
+        end
+
+        def definition_width(node)
+          node.loc.expression.source.squeeze.length -
+            node.loc.end.source.length
+        end
+
+        def max_line_length
+          config.for_cop('Metrics/LineLength')['Max'] || 80
+        end
+      end
+    end
+  end
+end